requirements.txt
asgiref==3.8.1
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.11
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
gunicorn==22.0.0
idna==3.7
isort==5.13.2
Markdown==3.6
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
psycopg2-binary==2.9.9
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
